+++ docs/changelog.md
@@ -10,6 +10,12 @@ and this project adheres to the
 [Python Version Specification](https://packaging.python.org/en/latest/specifications/version-specifiers/).
 See the [Contributing Guide](contributing.md) for details.
 
+## [Unreleased]
+
+### Fixed
+
+* Fix an issue with excessive backtracking when matching inline code blocks (#1617).
+
 ## [3.10.3] - 2026-07-30
 
 ### Fixed